Moved chain dropdown to its own component
