+++ src/ui/App.tsx
@@ -14,12 +14,13 @@ import {
   getNames, getStatus, Name, Status, SuccessResponse,
 } from '@sdk';
 import {
-  Layout, Select,
+  Layout,
   Typography,
 } from 'antd';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 
+import { ChainSelect } from '@components/ChainSelect';
 import { Console } from '@components/Console';
 import { MainMenu, MenuItems } from '@components/MainMenu';
 import { HelpPanel } from '@components/SidePanels/HelpPanel';
@@ -48,7 +49,7 @@ const useStyles = createUseStyles({
 });
 
 export const App = () => {
-  const { chain, setChain } = useGlobalState();
+  const { chain } = useGlobalState();
   dayjs.extend(relativeTime);
 
   const { setNamesMap } = useGlobalNames();
@@ -139,23 +140,6 @@ export const App = () => {
     },
   ];
 
-  // TODO: BOGUS - list of configured chains
-  const chainList = ['mainnet', 'gnosis', 'rinkeby', 'sepolia'];
-  const chainDropdown = (
-    <Select
-      placeholder='chain'
-      value={chain}
-      onChange={(newValue) => setChain(newValue)}
-      style={{ width: '10vw' }}
-    >
-      {chainList.map((item) => (
-        <Select.Option key={item} value={item}>
-          {item}
-        </Select.Option>
-      ))}
-    </Select>
-  );
-
   if (lastLocation && lastLocation !== '/' && location.pathname === '/') {
     return <Redirect to={lastLocation} />;
   }
@@ -167,7 +151,7 @@ export const App = () => {
         <Title style={{ color: 'white' }} level={2}>
           {`TrueBlocks Account Explorer - ${chain} chain`}
         </Title>
-        {chainDropdown}
+        <ChainSelect />
       </Header>
       <Layout>
         <SidePanel header='Menu' dir={PanelDirection.Left} cookieName='MENU_EXPANDED' collapsibleContent={false}>
